chore(node): move options files to `options` folder
