+++ packages/rolldown/src/options/watch-options.ts
@@ -1,5 +1,5 @@
-import { InputOptions } from '../types/input-options'
-import { OutputOptions } from '../types/output-options'
+import { InputOptions } from '../options/input-options'
+import { OutputOptions } from '../options/output-options'
 
 export interface WatchOptions extends InputOptions {
   output?: OutputOptions
