+++ packages/rolldown/src/rolldown-build.ts
@@ -1,8 +1,8 @@
 import { transformToRollupOutput } from './utils/transform-to-rollup-output'
 import { BundlerWithStopWorker, createBundler } from './utils/create-bundler'
 
-import type { InputOptions } from './types/input-options'
-import type { OutputOptions } from './types/output-options'
+import type { InputOptions } from './options/input-options'
+import type { OutputOptions } from './options/output-options'
 import type { RolldownOutput } from './types/rolldown-output'
 import type { HasProperty, TypeAssert } from './types/assert'
