+++ packages/rolldown/src/types/rolldown-options.ts
@@ -1,5 +1,5 @@
-import type { InputOptions } from '../types/input-options'
-import type { OutputOptions } from '../types/output-options'
+import type { InputOptions } from '../options/input-options'
+import type { OutputOptions } from '../options/output-options'
 
 export interface RolldownOptions extends InputOptions {
   // This is included for compatibility with config files but ignored by `rolldown.rolldown`
