+++ packages/rolldown/src/utils/normalize-plugin-option.ts
@@ -1,6 +1,6 @@
 import { asyncFlatten } from './async-flatten'
 import type { RolldownPlugin } from '../plugin'
-import type { InputOptions } from '../types/input-options'
+import type { InputOptions } from '../options/input-options'
 import type { OutputOptions, OutputPlugin } from '../rollup-types'
 
 export const normalizePluginOption: {
